COMMERCE-2769 - refactor, move classes to appropriate packages
