+++ modules/etl/mulesoft/src/main/java/com/liferay/mule/internal/connection/BaseCachedConnectionProvider.java
@@ -14,16 +14,24 @@
 
 package com.liferay.mule.internal.connection;
 
-import com.liferay.mule.internal.config.LiferayProxyConfig;
+import com.liferay.mule.internal.connection.config.LiferayProxyConfig;
 
 import java.io.IOException;
 
 import java.util.concurrent.TimeoutException;
 
+import javax.inject.Inject;
+
 import org.mule.runtime.api.connection.CachedConnectionProvider;
 import org.mule.runtime.api.connection.ConnectionException;
 import org.mule.runtime.api.connection.ConnectionValidationResult;
+import org.mule.runtime.api.meta.ExpressionSupport;
+import org.mule.runtime.extension.api.annotation.Expression;
+import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.param.ParameterGroup;
+import org.mule.runtime.extension.api.annotation.param.display.DisplayName;
+import org.mule.runtime.extension.api.annotation.param.display.Placement;
+import org.mule.runtime.http.api.HttpService;
 import org.mule.runtime.http.api.domain.message.response.HttpResponse;
 
 /**
@@ -63,7 +71,16 @@ public abstract class BaseCachedConnectionProvider
 		}
 	}
 
+	@Inject
+	protected HttpService httpService;
+
 	@ParameterGroup(name = "Proxy config")
 	protected LiferayProxyConfig liferayProxyConfig;
+
+	@DisplayName("OpenAPI Spec URL")
+	@Expression(ExpressionSupport.NOT_SUPPORTED)
+	@Parameter
+	@Placement(order = 1)
+	protected String openApiSpecPath;
 
 }
