+++ modules/etl/mulesoft/src/main/java/com/liferay/mule/internal/connection/BasicCachedConnectionProvider.java
@@ -14,19 +14,13 @@
 
 package com.liferay.mule.internal.connection;
 
-import com.liferay.mule.internal.config.BasicAuthenticationConfig;
-
-import javax.inject.Inject;
+import com.liferay.mule.internal.connection.config.BasicAuthenticationConfig;
 
 import org.mule.runtime.api.connection.ConnectionException;
-import org.mule.runtime.api.meta.ExpressionSupport;
 import org.mule.runtime.extension.api.annotation.Alias;
-import org.mule.runtime.extension.api.annotation.Expression;
-import org.mule.runtime.extension.api.annotation.param.Parameter;
 import org.mule.runtime.extension.api.annotation.param.ParameterGroup;
 import org.mule.runtime.extension.api.annotation.param.display.DisplayName;
 import org.mule.runtime.extension.api.annotation.param.display.Placement;
-import org.mule.runtime.http.api.HttpService;
 
 /**
  * @author Matija Petanjek
@@ -39,7 +33,7 @@ public class BasicCachedConnectionProvider
 	@Override
 	public LiferayConnection connect() throws ConnectionException {
 		return LiferayConnection.withBasicAuthentication(
-			_httpService, _openApiSpecPath,
+			httpService, openApiSpecPath,
 			_basicAuthenticationConfig.getUsername(),
 			_basicAuthenticationConfig.getPassword(),
 			liferayProxyConfig.getProxyConfig());
@@ -48,14 +42,5 @@ public class BasicCachedConnectionProvider
 	@ParameterGroup(name = "Connection config")
 	@Placement(order = 1)
 	private BasicAuthenticationConfig _basicAuthenticationConfig;
-
-	@Inject
-	private HttpService _httpService;
-
-	@DisplayName("OpenAPI Spec URL")
-	@Expression(ExpressionSupport.NOT_SUPPORTED)
-	@Parameter
-	@Placement(order = 2)
-	private String _openApiSpecPath;
 
 }
